Renamed 'Move Local Working Files Directory' to 'Move Local Corpus Storage Directory' so that it is clear to the user that the menu item will do. Modified the session storage file type to simple text files so that it is easer for administrators to maintain and to avoid issues with binary serialised java objects between jre versions. When the new format file does not exist the old format file will be read and a new format file created with those values. Prevented the locations list being saved unnecessarily. Enabled multiple delete of both remote corpus links and local directory links within the tree. Corrected some of the change working directory code to give better feedback to the user. Prevent corpus branches getting a template so that the global template takes effect. Added option to show the language id in the column title.

+++ src/nl/mpi/arbil/LinorgFavourites.java
@@ -29,18 +29,24 @@ public class LinorgFavourites {
     }
 
     // this will load any favourites in the old format and delete the old format file
-    public void loadOldFormatFavourites() {
+    public void convertOldFormatLocationLists() {
         try {
-            Vector<String> locationsList = (Vector<String>) LinorgSessionStorage.getSingleInstance().loadObject("locationsList");
+            File oldLocationsFile = new File(LinorgSessionStorage.getSingleInstance().storageDirectory, "locationsList");
             File oldFavouritesFile = new File(LinorgSessionStorage.getSingleInstance().storageDirectory, "selectedFavourites");
-            if (oldFavouritesFile.exists()) {
-                Vector<String> userFavouritesStrings = (Vector<String>) LinorgSessionStorage.getSingleInstance().loadObject("selectedFavourites");
-                locationsList.addAll(userFavouritesStrings);
-                LinorgSessionStorage.getSingleInstance().saveObject(locationsList, "locationsList");
-                oldFavouritesFile.deleteOnExit();
+            if (oldLocationsFile.exists()) {
+                Vector<String> locationsList = (Vector<String>) LinorgSessionStorage.getSingleInstance().loadObject("locationsList");
+                if (oldFavouritesFile.exists()) {
+                    Vector<String> userFavouritesStrings = (Vector<String>) LinorgSessionStorage.getSingleInstance().loadObject("selectedFavourites");
+                    locationsList.addAll(userFavouritesStrings);
+                    LinorgSessionStorage.getSingleInstance().saveStringArray("locationsList", locationsList.toArray(new String[]{}));
+                    oldFavouritesFile.deleteOnExit();
+                }
+                if (null == LinorgSessionStorage.getSingleInstance().loadStringArray("locationsList")) {
+                    LinorgSessionStorage.getSingleInstance().saveStringArray("locationsList", locationsList.toArray(new String[]{}));
+                }
             }
         } catch (Exception ex) {
-            System.out.println("load old format favourites failed: " + ex.getMessage());
+            GuiHelper.linorgBugCatcher.logError(ex);
         }
     }
